# core: avoid async op future reboxing to bundle PromiseId